Thunderplug: Huge revamp!
-Completely revamp hotplug start and exit to model module init and exit
-Flush workqueue on suspend
-Open up debug and hotplug_suspend to user
-Replace cpu suspending and resuming methods to methods similar to
msm_hotplug and intelli_hotplug
-Simplify a few checks
-Smoothen startup
-Tuning from dorimanx
-Turn off all cores when activity is stopped

Replaced cpus_online_all/offline_cpus method:
The methods were renamed to closer resemble each other, and a new
parameter is enabled to simplify code a bit. When initiating/exiting, we
want to bring up/down all CPUs, including CPU 1, while in suspend, we
want to keep CPU 1 awake to reduce lag and the possibility of a reboot.
A value of 1 means that CPU 1 will stay awake and a value of 0 means CPU
1 will be put to sleep.