Thunderplug: Additional reworking
Now it should truly support n-core CPUs, before you could set 4 max CPUs
on a dual-core system or not be able to set the max cores to 8 on an
octo-core system...silly. A switch is a nice option, however not when
the goal of the code is to be flexible.

Also force CPUs down during suspend, in order to act more similar to
msm_hotplug.